e kom kriju home controller me tregu a po funksionon porti

+++ src/main/java/com/example/flower_shop/repository/BouquetFlowerRepository.java
@@ -1,12 +1,11 @@
 package com.example.flower_shop.repository;
 
 import com.example.flower_shop.model.BouquetFlower;
-import com.example.flower_shop.model.Flower;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import java.util.List; 
 
 @Repository
-public interface BouquetFlowerRepository extends JpaRepository<BouquetFlower, Integer> {
-
+public interface BouquetFlowerRepository extends JpaRepository<BouquetFlower, Long> {
     List<BouquetFlower> findByBouquetId(Integer bouquetId);
 }
